FEAT: separate bottom-toolbar.js from common.js - bottom-toolbar.js is separted from common.js to add functions for bottom-toolbar - add a function to change favorite value '
